refactor: Add switchAlert function to chat API
Add a new function, switchAlert, to the chat API module. This function allows the user to switch the alert status of a chat room by making a POST request to the 'chat/room/alert' endpoint with the roomId and isAlert parameters. The response is logged to the console. If an error occurs, it is logged and thrown.

Co-authored-by: [Author Name]