```text refactor: Update dependencies and add StompJS and SockJS-Client
Update the dependencies in package.json to include the latest versions of React Navigation and other libraries. Additionally, add StompJS and SockJS-Client as new dependencies. This change improves the functionality and performance of the application by utilizing the latest versions of the libraries and adding support for STOMP messaging protocol.