refactor: Add ErrorIcon component to ChatDetail.styles.ts
The ChatDetail.styles.ts file in the src/screens/Chat directory has been modified to include a new ErrorIcon component. This component is used to display an error icon in the chat interface. The ErrorIcon component has a width and height of 20px and is styled using CSS. This refactor improves the visual representation of error messages in the chat interface.